feat(nuxt,docs): convert `<NuxtErrorBoundary>` to SFC; expose `error` and `clearError`
