+++ docs/3.api/1.components/6.nuxt-error-boundary.md
@@ -4,7 +4,7 @@ description: The <NuxtErrorBoundary> component handles client-side errors happen
 links:
   - label: Source
     icon: i-simple-icons-github
-    to: https://github.com/nuxt/nuxt/blob/main/packages/nuxt/src/app/components/nuxt-error-boundary.ts
+    to: https://github.com/nuxt/nuxt/blob/main/packages/nuxt/src/app/components/nuxt-error-boundary.vue
     size: xs
 ---
 
@@ -32,11 +32,34 @@ The `<NuxtErrorBoundary>` uses Vue's [`onErrorCaptured`](https://vuejs.org/api/c
     <template>
       <NuxtErrorBoundary>
         <!-- ... -->
-        <template #error="{ error }">
+        <template #error="{ error, clearError }">
           <p>An error occurred: {{ error }}</p>
+
+          <button @click="clearError">Clear error</button>
         </template>
       </NuxtErrorBoundary>
     </template>
   ```
 
 :read-more{to="/docs/getting-started/error-handling"}
+
+## Examples
+
+### Accessing `error` and `clearError` in script
+
+You can access `error` and `clearError` properties within the component's script as below:
+
+```vue
+<template>
+  <NuxtErrorBoundary ref="errorBoundary">
+    <!-- ... -->
+  </NuxtErrorBoundary>
+</template>
+
+<script setup lang="ts">
+const errorBoundary = useTemplateRef('errorBoundary')
+
+// errorBoundary.value?.error
+// errorBoundary.value?.clearError()
+</script>
+```
